Fix race condition in P1.memo()
The race was when two thread encountered the uninitialized state, `a == null`.
One thread gets suspended before entering the `synchronized` block, while the other runs though it. The latter thread sets `v` to something not null. Then the first thread resumes, enters the `synchronized` block and checks if `(v == null || v.get() == null)` which is now false, and thus returns the still uninitialized `a`.

Fixes #105